Move core elements in to core module of app.
