+++ core/src/main/java/com/hfad/palamarchuksuperapp/core/data/safeApiCall.kt
@@ -1,4 +1,4 @@
-package com.hfad.palamarchuksuperapp.data.services
+package com.hfad.palamarchuksuperapp.core.data
 
 import com.hfad.palamarchuksuperapp.domain.models.AppError
 import com.hfad.palamarchuksuperapp.domain.models.Result
