+++ core/src/main/java/com/hfad/palamarchuksuperapp/core/data/withSqlErrorHandling.kt
@@ -1,8 +1,8 @@
-package com.hfad.palamarchuksuperapp.data.repository
+package com.hfad.palamarchuksuperapp.core.data
 
 import android.database.SQLException
-import com.hfad.palamarchuksuperapp.domain.models.AppError
-import com.hfad.palamarchuksuperapp.domain.models.Result
+import com.hfad.palamarchuksuperapp.core.domain.AppError
+import com.hfad.palamarchuksuperapp.core.domain.Result
 
 suspend fun <T> withSqlErrorHandling(
     block: suspend () -> T,
