+++ core/src/main/java/com/hfad/palamarchuksuperapp/core/domain/Result.kt
@@ -1,4 +1,4 @@
-package com.hfad.palamarchuksuperapp.domain.models
+package com.hfad.palamarchuksuperapp.core.domain
 
 typealias RootError = Error
