+++ core/src/main/java/com/hfad/palamarchuksuperapp/core/ui/composables/basic/AppText.kt
@@ -1,4 +1,4 @@
-package com.hfad.palamarchuksuperapp.ui.reusable.elements
+package com.hfad.palamarchuksuperapp.core.ui.composables.basic
 
 import androidx.annotation.StringRes
 import androidx.compose.foundation.background
