+++ core/src/main/java/com/hfad/palamarchuksuperapp/core/ui/composables/effects/pulseEffect.kt
@@ -1,4 +1,4 @@
-package com.hfad.palamarchuksuperapp.ui.reusable
+package com.hfad.palamarchuksuperapp.core.ui.composables.effects
 
 import androidx.compose.animation.core.DurationBasedAnimationSpec
 import androidx.compose.animation.core.FastOutSlowInEasing
