+++ core/src/main/java/com/hfad/palamarchuksuperapp/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.hfad.palamarchuksuperapp.ui.compose.theme
+package com.hfad.palamarchuksuperapp.core.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
